% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eg_operationaldata.R
\name{eg_agg}
\alias{eg_agg}
\title{Get aggregate data.}
\usage{
eg_agg(
  indicator = NULL,
  periodType = NULL,
  from = NULL,
  to = NULL,
  countryKey = NULL,
  countryLabel = NULL,
  operatorKey = NULL,
  tsoEicCode = NULL,
  operatorLabel = NULL,
  pointKey = NULL,
  pointLabel = NULL,
  tsoItemIdentifier = NULL,
  directionKey = NULL,
  timeZone = "CET",
  limit = -1
)
}
\arguments{
\item{indicator}{Indicator}

\item{periodType}{Period type.}

\item{countryKey}{Country key. BE, DE, DK etc.}

\item{countryLabel}{Country Label. Germany, Belgium, etc.}

\item{operatorKey}{Operator key.}

\item{tsoEicCode}{TSO EIC code.}

\item{operatorLabel}{Operator Label.}

\item{pointKey}{Point Key.}

\item{pointLabel}{Point Labe.}

\item{tsoItemIdentifier}{TSO Item Identifier.}

\item{directionKey}{Direction Key.}

\item{limit}{Limit. -1 for unlimited.}

\item{periodFrom}{Period from.}

\item{periodTo}{Period to.}

\item{timezone}{Time zone.}
}
\description{
All the connections between transmission system
operators and their respective balancing zones
}
