% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eg_operationaldata.R
\name{eg_connectionpoints}
\alias{eg_connectionpoints}
\title{Get interconnection data.}
\usage{
eg_connectionpoints(indicator = NULL, periodType = NULL, from = NULL,
  to = NULL, operatorKey = NULL, tsoEicCode = NULL,
  operatorLabel = NULL, pointKey = NULL, pointLabel = NULL,
  tsoItemIdentifier = NULL, directionKey = NULL, timeZone = "CET",
  limit = -1)
}
\arguments{
\item{indicator}{Indicator}

\item{periodType}{Period type.}

\item{operatorKey}{Operator key.}

\item{tsoEicCode}{TSO EIC code.}

\item{operatorLabel}{Operator Label.}

\item{pointKey}{Point Key.}

\item{pointLabel}{Point Labe.}

\item{tsoItemIdentifier}{TSO Item Identifier.}

\item{directionKey}{Direction Key.}

\item{limit}{Limit. -1 for unlimited.}

\item{periodFrom}{Period from.}

\item{periodTo}{Period to.}

\item{timezone}{Time zone.}
}
\description{
Interconnection points
}
