% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eg_operationaldata.R
\name{eg_operationaldata_physical_flow}
\alias{eg_operationaldata_physical_flow}
\title{Get interconnections data.}
\usage{
eg_operationaldata_physical_flow(pointDirection, from, to,
  periodType = "hour", datetime_utc_convert = TRUE,
  remove_forward_dt = TRUE, replace_known_operator_zeros = TRUE)
}
\arguments{
\item{pointDirection}{PointDirection from entsog.}

\item{from}{From as.Date.}

\item{to}{To. as.Date.}

\item{periodType}{PeriodType. Defaults to 'hour'.}

\item{datetime_utc_convert}{Convert datetimes to UTC. Defaults to TRUE.}

\item{remove_forward_dt}{Remove datetime observations ahead in time. This is a problem for some
variables where there shouldn't be observations as they are only historic.}

\item{replace_known_operator_zeros}{Some operators give zeros instead of NAs, for ex NET4GAS.
The zeros are replaced with NAs.}
}
\description{
Get interconnections data.
}
\examples{

library(entsog)
library(tidyverse)

nordstream <- c("cz-tso-0001itp-00010entry", "de-tso-0016itp-00251entry", "de-tso-0017itp-00247entry")
opal <- c("de-tso-0016itp-00251entry", "de-tso-0017itp-00247entry", "cz-tso-0001itp-00010entry")

head(eg_operationaldata_physical_flow(opal, as.Date("2019-10-14"), as.Date("2019-10-15")))

}
