% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eg_operatorpointdirections.R
\name{eg_operatorpointdirection}
\alias{eg_operatorpointdirection}
\title{Get operator point directions data.}
\usage{
eg_operatorpointdirection()
}
\description{
Get operator point directions data.
}
